Support multiple concurrent SQLite databases
Uses a `DataSet` abstraction as a base class of a SQLAlchemy declarative base class
to define unifying functionality for related mappings. The SQLite configuration
defaults to in-memory, but can be overridden per data set name using standard microcosm
configuration or entry points (e.g. for libraries that define a single data set).