Avoid configuring DB paths via entry points in tests
Having the DB path configuration via an entry point allows a library user to load sqlite data without worrying about where the DB file is.
In unit tests however, it means that if we've defined an entry point to configure the path, we'll have to explicitly set SQLite to use an in-memory DB in the test config.
This is somehwat dangerous, since a test written by someone ignoring this could modify the actual DB that the library exports.